==> 1.Assignation.ipynb <==
# ASSIGNATIONS TAXONOMIQUES

## CHARGEMENT DES PACKAGES

library(Rcpp)
library(dada2)
library(Biostrings)
library(tidyverse) 

## OUVERTURE DES SEQUENCES

# ouverture du fichier .fasta 
seq <- readDNAStringSet("data/sequences.fasta", format="fasta",
               nrec=-1L, skip=0L, seek.first.rec=FALSE,
               use.names=TRUE, with.qualities=FALSE)
seq

# Transformation des desquences en vecteurs de charactères, necessaire pour la fonction assignTaxonomy
seq_char <- as.character(seq)
head(seq_char)

## ASSIGNATION TAXONOMIQUE SELON LES DATABASES

## 1) RDP

# Assignation taxonomique en utilisant la base de référence RDP
taxa <- assignTaxonomy(seq_char,"data/TaxoRef_Training/rdp_19_toSpecies_trainset.fa.gz", multithread=TRUE)
head(taxa)

# Convertir la matrice taxa en data frame
taxa_RDP <- as.data.frame(taxa)
taxa_RDP <- rownames_to_column(taxa_RDP, var = "sequence")
                                 
# Ajouter une colonne "ASV" avec les noms de seq_char
taxa_RDP$ASV <- names(seq_char)

# Ajouter une colonne "base_ref" avec la valeur "RDP" pour toutes les lignes
taxa_RDP$base_ref <- "RDP"

# supprimer la colonne séquence

taxa_RDP$sequence <- NULL
head(taxa_RDP)

EXPORTATION TABLEAU

write.table(taxa_table, file= "data/output/assign_RDP.csv", sep= ";" , row.names= FALSE, col.names= TRUE)

## 2) SILVA

# Assignation taxonomique en utilisant la base de référence Silva
taxa <- assignTaxonomy(seq_char, "data/TaxoRef_Training/silva_nr99_v138.2_toSpecies_trainset.fa.gz", multithread=TRUE)
head(taxa)

# Convertir la matrice taxa en data frame
taxa_SILVA <- as.data.frame(taxa)
taxa_SILVA <- rownames_to_column(taxa_SILVA, var = "sequence")
                                 
# Ajouter une colonne avec les noms de seq_char
taxa_SILVA$ASV <- names(seq_char)

# Ajouter une colonne "base_ref" avec la valeur "SILVA" pour toutes les lignes
taxa_SILVA$base_ref <- "SILVA"

# supprimer la colonnes séquence

taxa_SILVA$sequence <- NULL
head(taxa_SILVA)

EXPORTATION TABLEAU

write.table(taxa_SILVA, file= "data/output/assign_SILVA.csv", sep= ";" , row.names= FALSE, col.names= TRUE)

## 3) GG2

# Assignation taxonomique en utilisant la base de référence GreenGenes
taxa <- assignTaxonomy(seq_char, "data/TaxoRef_Training/gg2_2024_09_toSpecies_trainset.fa.gz", multithread=TRUE)
head(taxa)

# Convertir la matrice taxa en data frame
taxa_gg2 <- as.data.frame(taxa)
taxa_gg2 <- rownames_to_column(taxa_gg2, var = "sequence")
                                 
# Ajouter une colonne avec les noms de seq_char
taxa_gg2$ASV <- names(seq_char)

# Ajouter une colonne "base_ref" avec la valeur "GG2" pour toutes les lignes
taxa_gg2$base_ref <- "GG2"

# supprimer la colonnes séquence

taxa_gg2$sequence <- NULL
head(taxa_gg2)

# Retirer le "préfixe" au niveau des rangs taxonomiques : 
taxa_gg2 <- taxa_gg2 |> 
    # modifier toutes les chaines de caractères commençant par une lettre en minuscule, suivie d'un double 
    # underscore pour les retirer et ne conserver que ce qui suit
    mutate(across(everything(), ~ str_remove(., "^[a-z]__")))

EXPORTATION TABLEAU 

write.table(taxa_gg2, file= "data/output/assign_gg2.csv", sep= ";" , row.names= FALSE, col.names= TRUE)
