==> 2.Preparation_data.ipynb <==
# Preparation du tableau de données pour les analyses statistiques

## Choix du dossier de travail

# Donner le chemin vers le dossier de travail choisi
setwd("shared/projects/sorbonne_master_ue_projet/Colaris")

## Installation des packages

# Installation du package qiime2R depuis github s'il n'est pas déjà présent
if (!requireNamespace("devtools", quietly = TRUE)){install.packages("devtools")}
devtools::install_github("jbisanz/qiime2R")

### Chargement des librairies

library(qiime2R)
library(tidyverse)

### Importation des données

#### Metadonnées de l'expérience

# Importation des meta données associées à l'expérience sur la température et le nourrissage de 4 espèces de nématodes
meta_data <-read.table("data/meta-data2.tsv", header= T)

#### Table de contingence des ASVs

# Importation de la table de contingence des ASVs de chaque échantillon
ASV <- read_qza("data/table-nematodesAM.qza")

# Création d'une table comprenant uniquement les données sur les ASVs et leur nombre
ASV_tbl <- as.data.frame(ASV$data) 

# Transposer le dataframe pour avoir un échantillon par ligne et les ASVs en colonne
ASV_tbl <- as.data.frame(t(ASV_tbl)) 

# Transformer le nom des lignes en une colonne intitulée "sample.id"
ASV_tbl <- rownames_to_column(ASV_tbl, var = "sample.id") 

#### Table des assignations taxonomiques en fonction des bases de référence

Dans cette partie, nous regroupons les assignations taxonomiques obtenues dans la partie [1.Assignation.ipynb](/1.Assignation.ipynb). 

# Importation des assignations taxonomique
assign_RDP <- tibble(read.csv("data/output/assign_RDP.csv",sep = ";"))
assign_SILVA <- tibble(read.csv("data/output/assign_SILVA.csv",sep = ";"))
assign_GG2 <- tibble(read.csv("data/output/assign_GG2.csv",sep = ";"))

# Regrouper les trois tables en un seul dataframe en conservant toutes les lignes de chaque dataframe
assign_3ref <- bind_rows(assign_RDP, assign_SILVA, assign_GG2)

### Fusion des dataframes

# Fusionner la table de metadonnées et la table de contingence des ASVs en fonction de l'échantillon
PM4_data <- merge(ASV_tbl,meta_data, by= "sample.id") |> filter(species== "PM4" & food == "with_food")|> select(-c('species','food','replicats','treatment','description'))

# Pivoter le tableau resultant en version longue pour avoir un ASV par ligne pour tous les échantillons
PM4_data_lg <- pivot_longer(PM4_data, cols=-c(sample.id,temperature),names_to ="ASV",values_to ="frequence") |> 
    filter(frequence > 0) # filtrer pour ne garder que les lignes pour lesquelles la fréquence de l'ASV est supérieure à 0 

# Fusionner PM4_data_lg avec les assignations
data_final_pm4 <- merge(PM4_data_lg,assign_3ref, by= "ASV")

## Sauvegarde des jeux de données crées

# Sauvegarder la fusion des métadonnées et des contingence d'ASVs : PM4_data_lg
write.table(PM4_data_lg, file = "data/output/PM4_data_clean.csv", sep = ";",row.names = FALSE,col.names = TRUE)

# Sauvegarder le tableau de données final regroupant assignations, metadonnées et contingence d'ASVs
write.table(data_final_pm4, file = "data/output/data_final_pm4.csv", sep = ";", row.names = FALSE, col.names = TRUE)
