==> 4.Dominance.ipynb <==
# DOMINANCE ET E.coli

# Chargement des librairies
library(tidyverse)
library(dplyr)

# Importation du tableau de donnée final
df_pm4 <- tibble(read.csv("data/output/data_final_pm4.csv", sep = ";"))
head(df_pm4)

# I- VERIFICATION DU NOMBRE DE E.coli DANS LES DATABASES

Pour assurer la fiabilité des analyses, nous avons d'abord vérifié l'abondance de E. coli car les nématodes ont été nourris avec cette bactérie.

nb_coli <- df_pm4 %>%
  filter(Species == "coli") %>%       # Filtrer uniquement les lignes où l'espèce est "coli"
  group_by(base_ref) %>%              # Regrouper par base de reference
  dplyr::summarise(
    nb_asv_distincts = n_distinct(ASV),  # Calcul du nombre d'ASV distincts
    nb_coli = n()                        # Calcul du nombre total d'occurrences de "coli"
  )

# Afficher les résultats
print(nb_coli)

# II- CALCUL DU GENRE ET DE L'ESPECE DOMINANTE 

Nous avons analysé les données pour identifier l’espèce et le genre les plus abondants en fonction des différentes températures et des bases de données utilisées.

## 1) L'espece dominante 

# Trouver l'espèce la plus abondante pour chaque combinaison de base de données et température
resultat <- df_pm4 %>%
 group_by(temperature) %>% # grouper les températures 
filter(frequence == max(frequence)) %>% # filtrer les lignes avec la fréquence la plus élevée
  group_by(base_ref) %>% # grouper par base de references 
  ungroup() %>% # dégrouper 
  select(base_ref, temperature, Species, frequence)  # Sélectionner les colonnes importantes

# Afficher les résultats
print(resultat)

## 2) Le genre dominant

# Trouver le genre le plus abondant pour chaque combinaison de base de donnée et température
resultat <- df_pm4 %>%
  group_by(base_ref, temperature) %>% # grouper par base de references et temperature 
  filter(frequence == max(frequence)) %>% # filtrer les lignes avec la fréquence la plus élevée
  ungroup() %>% # dégrouper 
  select(base_ref, temperature, Genus, frequence)  # Sélectionner les colonnes importantes

# Afficher les résultats
print(resultat)
